+++ app/code/Firebear/ImportExport/Model/Source/Type/Rest.php
@@ -6,6 +6,7 @@
 
 namespace Firebear\ImportExport\Model\Source\Type;
 
+use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Filesystem\DriverPool;
 
 /**
@@ -15,9 +16,6 @@ use Magento\Framework\Filesystem\DriverPool;
  */
 class Rest extends AbstractType
 {
-    const JSON_FILENAME_EXTENSION = '.json';
-    const XML_FILENAME_EXTENSION = '.xml';
-
     /**
      * @var string
      */
@@ -31,7 +29,7 @@ class Rest extends AbstractType
     /**
      * Download remote source file to temporary directory
      *
-     * @return bool|strxing
+     * @return bool|string
      * @throws \Magento\Framework\Exception\LocalizedException
      */
     public function uploadSource()
@@ -46,7 +44,7 @@ class Rest extends AbstractType
                         mkdir($dirname, 0775, true);
                     }
                 } catch (\Exception $e) {
-                    throw new  \Magento\Framework\Exception\LocalizedException(
+                    throw new  LocalizedException(
                         __(
                             "Can't create local file /var/import/json'. Please check files permissions. "
                             . $e->getMessage()
@@ -57,61 +55,90 @@ class Rest extends AbstractType
                 $fileMetadata = $client->execute(
                     $this->getData('request_url'),
                     $this->getData('request_method'),
-                    $this->convertToFormat($this->getData('request_body'))
+                    $this->convertToFormat($this->getData('request_body')),
+                    $this->getHeaders()
                 );
                 if ($fileMetadata && false !== $fileMetadata->response) {
                     file_put_contents($filePath, $fileMetadata->response);
                     $this->fileName = $this->getImportPath() . '/' . $fileName;
                 } else {
-                    throw new \Magento\Framework\Exception\LocalizedException(__("No content from API call, error: " . $fileMetadata->error));
+                    throw new LocalizedException(__('No content from API call, error: ' . $fileMetadata->error));
                 }
             }
 
             return $this->fileName;
         } else {
-            throw new  \Magento\Framework\Exception\LocalizedException(__("Can't initialize %s client", $this->code));
+            throw new  LocalizedException(__("Can't initialize %s client", $this->code));
         }
     }
 
     /**
-     * Download remote images to temporary media directory
-     *
      * @param $importImage
      * @param $imageSting
      *
-     * @return bool
+     * @return array
      */
     public function importImage($importImage, $imageSting)
     {
+        $imageExt = '.jpeg';
         $filePath = $this->directory->getAbsolutePath($this->getMediaImportPath() . $imageSting);
-        $dirname = dirname($filePath);
-        if (!is_dir($dirname)) {
-            mkdir($dirname, 0775, true);
+        $dirname = \dirname($filePath);
+        if (!mkdir($dirname, 0775, true) && !is_dir($dirname)) {
+            throw new \RuntimeException(sprintf('Directory "%s" was not created', $dirname));
         }
+
         if (preg_match('/\bhttps?:\/\//i', $importImage, $matches)) {
             $url = str_replace($matches[0], '', $importImage);
         } else {
             $sourceFilePath = $this->getData($this->code . '_file_path');
-            $sourceDir = dirname($sourceFilePath);
+            $sourceDir = \dirname($sourceFilePath);
             $url = $sourceDir . '/' . $importImage;
             if (preg_match('/\bhttps?:\/\//i', $url, $matches)) {
                 $url = str_replace($matches[0], '', $url);
             }
         }
         if ($url) {
+            $fileExistOnServer = 0;
             try {
                 $driver = $this->getProperDriverCode($matches);
                 $read = $this->readFactory->create($url, $driver);
-                $this->directory->writeFile(
-                    $filePath,
-                    $read->readAll()
-                );
+                $fileExistOnServer = 1;
             } catch (\Exception $e) {
-
+            }
+            if ($fileExistOnServer) {
+                try {
+                    $this->directory->writeFile(
+                        $filePath,
+                        $read->readAll()
+                    );
+                    if (\function_exists('mime_content_type')) {
+                        $imageExt = $this->getImageMimeType(mime_content_type($filePath));
+                        $this->directory->renameFile(
+                            $filePath,
+                            $filePath . $imageExt
+                        );
+                    }
+                } catch (\Exception $e) {
+                }
             }
         }
 
-        return true;
+        return [true, $imageSting . $imageExt];
+    }
+
+    protected function getImageMimeType($mime)
+    {
+        switch ($mime) {
+            case 'image/jpeg':
+                $ext = '.jpeg';
+                break;
+            case 'image/png':
+                $ext = '.png';
+                break;
+            default:
+                $ext = '.jpeg';
+        }
+        return $ext;
     }
 
     public function importImageCategory($importImage, $imageSting)
@@ -140,7 +167,6 @@ class Rest extends AbstractType
                     $read->readAll()
                 );
             } catch (\Exception $e) {
-
             }
         }
 
@@ -176,26 +202,27 @@ class Rest extends AbstractType
     }
 
     /**
-     * @param string      $data Json request body from Job's settings
+     * @param string $data Json request body from Job's settings
      * @param null|string $type
-     * @param null        $format
+     * @param null $format
      *
      * @return mixed json-structured data
      */
     public function convertToFormat($data, $type = null, $format = null)
     {
         $result = '';
-        $type = $type ? $type : $this->getData('type_file');
+        $type = $type ?: $this->getData('type_file');
         switch ($type) {
             case 'xml':
-//                $data = trim(preg_replace('/\s+/', '', $data));
                 $result = $data;
                 break;
             case 'json':
-                $data = trim(preg_replace('/\s+/', '', $data));
-                if ($format == 'object'){
+                if ($format == 'object') {
                     $result = json_decode($data, true);
-                } else{
+                    if (!$result) {
+                        $result = [];
+                    }
+                } else {
                     $result = json_encode(json_decode($data, true));
                 }
                 break;
@@ -203,22 +230,85 @@ class Rest extends AbstractType
                 // ToDo: make an exception and catch it further
                 break;
         }
-
         return $result;
     }
 
     /**
-     * @param string $url
+     * @param $matches
      *
      * @return string
      */
-    public function convertUrlToFilename($url)
+    protected function getProperDriverCode($matches)
     {
-        $parsedUrl = parse_url($url);
-        $filename = str_replace('.', '_', $parsedUrl['host'])
-            . str_replace('/', '_', $parsedUrl['path'])
-            . constant("self::" . strtoupper($this->getData('type_file')) . "_FILENAME_EXTENSION");
+        if (is_array($matches)) {
+            return (false === strpos($matches[0], 'https'))
+                ? DriverPool::HTTP
+                : DriverPool::HTTPS;
+        } else {
+            return DriverPool::HTTP;
+        }
+    }
 
-        return $filename;
+    /**
+     * @param \Firebear\ImportExport\Model\Export $model
+     *
+     * @return array
+     */
+    public function run(\Firebear\ImportExport\Model\Export $model)
+    {
+        $response = '';
+        $result = true;
+        $errors = [];
+        $path = '';
+        try {
+            $fileFormat = $model->getFileFormat();
+            $this->setExportModel($model);
+            $name = 'export_' . $this->timezone->date()->format('Y_m_d_H_i_s') .
+                '.' . $fileFormat;
+            $path = AbstractType::EXPORT_DIR . '/' . $name;
+            if ($this->writeFile($path)) {
+                if ($client = $this->_getSourceClient()) {
+                    $this->setData('type_file', $fileFormat);
+                    $currentDate = '';
+                    if ($this->getData('date_format')) {
+                        $format = $this->getData('date_format') ?? 'Y-m-d-hi';
+                        $currentDate = '-' . $this->timezone->date()->format($format);
+                    }
+
+                    $filePath = $this->directory->getAbsolutePath($path);
+                    $dataRead = $this->readFactory->create($filePath, DriverPool::FILE);
+                    $headers = $this->getHeaders();
+                    if ($this->getData('request_method') === 'DELETE') {
+                        $result = $client->delete($this->getData('request_url'), $dataRead->readAll(), $headers);
+                    } else {
+                        $result = $client->put($this->getData('request_url'), $dataRead->readAll(), $headers);
+                    }
+                    $response = $result->response;
+                    $result = true;
+                    if (!$result) {
+                        $result = false;
+                        $errors[] = __('File not found');
+                    }
+                } else {
+                    $result = false;
+                    $errors[] = __("Can't initialize %s client", $this->code);
+                }
+            } else {
+                $result = false;
+            }
+        } catch (\Exception $e) {
+            $result = false;
+            $errors[] = __('Folder for import / export don\'t have enough permissions! Please set 775');
+        }
+
+        return [$result, $path, $errors, $response];
+    }
+
+    /**
+     * @return array
+     */
+    public function getHeaders()
+    {
+        return [];
     }
 }
